Move models to their own packages + implement getAllUsers
