+++ app/src/androidTest/java/com/github/cmput301f21t44/hellohabits/HabitEventTest.java
@@ -19,7 +19,7 @@ import androidx.test.filters.LargeTest;
 
 import com.github.cmput301f21t44.hellohabits.firebase.FSHabit;
 import com.github.cmput301f21t44.hellohabits.firebase.FSUser;
-import com.github.cmput301f21t44.hellohabits.model.DaysOfWeek;
+import com.github.cmput301f21t44.hellohabits.model.habit.DaysOfWeek;
 import com.github.cmput301f21t44.hellohabits.view.MainActivity;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.firestore.FirebaseFirestore;
