+++ app/src/main/java/com/github/cmput301f21t44/hellohabits/model/habit/DaysOfWeek.java
@@ -1,4 +1,4 @@
-package com.github.cmput301f21t44.hellohabits.model;
+package com.github.cmput301f21t44.hellohabits.model.habit;
 
 import androidx.annotation.NonNull;
