+++ app/src/main/java/com/github/cmput301f21t44/hellohabits/model/social/Follow.java
@@ -1,4 +1,4 @@
-package com.github.cmput301f21t44.hellohabits.model;
+package com.github.cmput301f21t44.hellohabits.model.social;
 
 import java.util.Collections;
 import java.util.Map;
