+++ app/src/main/java/com/github/cmput301f21t44/hellohabits/view/habit/HabitAdapter.java
@@ -13,7 +13,7 @@ import androidx.recyclerview.widget.ListAdapter;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.github.cmput301f21t44.hellohabits.databinding.ListHabitItemBinding;
-import com.github.cmput301f21t44.hellohabits.model.Habit;
+import com.github.cmput301f21t44.hellohabits.model.habit.Habit;
 import com.github.cmput301f21t44.hellohabits.view.OnItemClickListener;
 
 /**
