+++ app/src/main/java/com/github/cmput301f21t44/hellohabits/view/habitevent/HabitEventAdapter.java
@@ -9,7 +9,7 @@ import androidx.recyclerview.widget.ListAdapter;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.github.cmput301f21t44.hellohabits.databinding.ListHabitEventItemBinding;
-import com.github.cmput301f21t44.hellohabits.model.HabitEvent;
+import com.github.cmput301f21t44.hellohabits.model.habitevent.HabitEvent;
 import com.github.cmput301f21t44.hellohabits.view.OnItemClickListener;
 
 import java.time.ZoneId;
